requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
httpx>=0.27
pydantic>=2.6
pandas>=2.1
openpyxl>=3.1
pypdf>=4.2
python-docx>=1.1
python-dotenv>=1.0
